Clean up broken result.
